# specification-module/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Store API - OpenAPI 3.0
  description: |-
    This is a Store API Server
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://store-api.swagger.io/api/v3
...
paths:
...
  /inventoryItem:
    put:
      tags:
        - inventoryItem
      summary: Update an existing Inventory Item
      description: Update an existing Branch by Id
      operationId: updateInventoryItem
      requestBody:
        description: Update an existent Inventory Item in the app
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItemDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/InventoryItemDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/InventoryItemDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/InventoryItemResponseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Inventory Item not found
        '405':
          description: Validation exception
      security:
        - bearerAuth: []
    post:
      tags:
        - inventoryItem
      summary: Add a new Inventory Item to the app
      description: Add a new Inventory Item to the app
      operationId: addInventoryItem
      requestBody:
        description: Create a new Inventory Item to the app
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItemDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/InventoryItemDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/InventoryItemDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/InventoryItemResponseDto'
        '405':
          description: Invalid input
      security:
        - bearerAuth: []
  /inventoryItem/findByStatus:
    get:
      tags:
        - branch
      summary: Finds Inventory Item by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findInventoryItemByStatus
      parameters:
        - name: status
          in: query
          description: Status values that need to be considered for filter
          required: false
          explode: true
          schema:
            type: string
            default: ACTIVE
            enum:
              - ACTIVE
              - DEACTIVE
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItemResponseDto'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryItemResponseDto'
        '400':
          description: Invalid status value
      security:
        - bearerAuth: []
  /inventoryItem/{inventoryItemId}:
    get:
      tags:
        - inventoryItem
      summary: Find Inventory Item by ID
      description: Returns a single Inventory Item
      operationId: getInventoryItemById
      parameters:
        - name: inventoryItemId
          in: path
          description: ID of Inventory Item to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItemResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/InventoryItemResponseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Inventory Item not found
      security:
        - bearerAuth: []
    post:
      tags:
        - inventoryItem
      summary: Updates a Inventory Item in the app with form data
      description: ''
      operationId: updateInventoryItemWithForm
      parameters:
        - name: inventoryItemId
          in: path
          description: ID of Inventory Item to return
          required: true
          schema:
            type: integer
            format: int64
        - name: branchId
          in: query
          description: Branch ID of Inventory Item that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: batch_id
          in: query
          description: Store ID of Inventory Item that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: quantity
          in: query
          description: Quantity of Inventory Item that needs to be updated
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          description: Status of Branch that needs to be updated
          schema:
            type: string
      responses:
        '405':
          description: Invalid input
      security:
        - bearerAuth: []
    delete:
      tags:
        - inventoryItem
      summary: Deletes a Inventory Item
      description: delete a Inventory Item
      operationId: deleteInventoryItem
      parameters:
        - name: api_key
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: inventoryItemId
          in: path
          description: Inventory Item id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid Inventory Item value
      security:
        - bearerAuth: []
components:
  schemas:
...
    BatchDto:
      required:
        - batchName
        - sellingPrice
        - skuId
      type: object
      properties:
        batchId:
          type: integer
          format: int64
          example: 10
        batchName:
          type: string
          example: 1st batch, 2nd batch
        buyingPrice:
          type: number
          format: double
          example: 100.00
        sellingPrice:
          type: number
          format: double
          example: 100.00
        photoUrl:
          type: string
        skuId:
          type: integer
          format: int64
          example: 10
        status:
          type: string
          description: Batch status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: batch
...
    AddressDto:
      type: object
      properties:
        addressId:
          type: integer
          format: int64
          example: 10
        addresLine1:
          type: string
          example: 437 Lytton
        addresLine2:
          type: string
          example: 437 Lytton
        city:
          type: string
          example: Palo Alto
        state:
          type: string
          example: CA
        zip:
          type: string
          example: '94301'
        country:
          type: string
          example: Canada
      xml:
        name: AddressDto
    BranchDto:
      required:
        - branchName
        - storeId
      type: object
      properties:
        branchId:
          type: integer
          format: int64
          example: 10
        storeId:
          type: integer
          format: int64
          example: 10
        branchName:
          type: string
          example: Branch Toronto
        address:
          $ref: '#/components/schemas/AddressDto'
        photoUrl:
          type: string
        status:
          type: string
          description: Branch status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: BranchDto
...
    InventoryItemDto:
      required:
        - branchId
        - batchId
      type: object
      properties:
        inventoryItemId:
          type: integer
          format: int64
          example: 10
        branchId:
          type: integer
          format: int64
          example: 10
        batchId:
          type: integer
          format: int64
          example: 10
        quantity:
          type: integer
          format: int64
          example: 10
        status:
          type: string
          description: Branch status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: inventoryItem
    InventoryItemResponseDto:
      required:
        - branchId
        - batchId
      type: object
      properties:
        inventoryItemId:
          type: integer
          format: int64
          example: 10
        branch:
          $ref: '#/components/schemas/BranchDto'
        batch:
          $ref: '#/components/schemas/BatchDto'
        quantity:
          type: integer
          format: int64
          example: 10
        status:
          type: string
          description: Branch status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: InventoryItemResponseDto
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT